deploy/requirements.txt
Fabric3==1.14.post1
click==6.7
